requirements.txt
apache-airflow-providers-snowflake
snowflake-connector-python
apache-airflow-providers-common-sql
snowflake-sqlalchemy
